EPAS 15 TDE vs no-TDE benchmark medium t-shirt amd
A copy of aws-tpcc-tde until we parameterize instance types and other
configuration variables.

This is the TDE vs no-TDE benchmark implementation.
As EPAS 15 packages are still in beta version, we are using
CloudSmith beta repo. This must to be changed once EPAS 15 is GA.

Environment variables:
- BENCHMARK_NAME - Name of the benchmark, used to store relative
data on the S3 bucket.
- TERRAFORM_PROJECT_PATH - Terraform project path.
- TPCC_WAREHOUSE - Number of warehouses.
- TPCC_DURATION - Long run duration, in minutes.
- TPCC_RAMPUP - Rampup duration in long run, in minutes.
- TPCC_VUSERS - Number of virtual users for the long run.
- EDB_BETA_REPO_TOKEN - Access token to EDB's beta repos