Dirty hack for tde medium amd test
Dirty hack in the interest of time.  Let the post-processing.py script
process what it can and fix the broken parts later.  This also let's us
not worry about whether all the files exists for the push_results steps,
meaning we don't have to hack that for the time being.

Then we can manually analyze the data collected.