Added role to introduce latency on selected hosts.
This adds the post-provision role for general use within all
benchmarks. Within this role is the add-latency list of tasks.
This task list is only included if the latency_ms variable is
non-zero. It is also defined as 0 in the list of default variables
to avoid errors from including this module and not setting the
variable. It should be a drop-in addition to any benchmark that
includes the role for all systems.